word_counter_dsc/requirements.txt
discord.py>=2.4.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1

# Optional (only needed if DB_DIALECT=postgres)
asyncpg>=0.29.0
